修复多线程下载线程的问题，并且改变所有的Int为Long 新增自定义输入法键盘

+++ edge/src/main/java/com/daniel/edge/management/download/model/DownloadModel.kt
@@ -3,6 +3,6 @@ package com.daniel.edge.management.download.model
 /**
  * @param state -1删除，0暂停，1开始
  */
-data class DownloadModel(var id:Int,var name: String, var url: String,var state:Int = 0,var currentTime:Long) {
+data class DownloadModel(var id:Long,var name: String, var url: String,var state:Int = 0,var currentTime:Long) {
     var threadModels: ArrayList<DownloadThreadModel>? = null
 }
